[CMake] Make INSTALL_IN_COMPONENT optional
Several CMake functions (e.g. add_swift_target_library) take the option
`INSTALL_IN_COMPONENT`, specifying which swift component you want a target to be
associated with for the purposes of installation. However, there are several
instances where you want to add a library but never install them. The workaround
for this was to make the swift component `never_install`. Although this is kind
of a hack, it does work. However, I am in the process of adding CMake
targets for the components themselves so they can have installation targets.
Instead of propagating the `never_install` workaround, I think it would be
cleaner to make `INSTALL_IN_COMPONENT` optional and remove `never_install`
completely.