ASTPrinter: Fix generic signatures with requirements on outer parameters
We incorrectly assumed that a requirement 'rooted' in an outer parameter
must be necessarily defined as part of the outer context's signature,
and thus we were skipping it when printing the 'where' clause of a
nested declaration or an extension of a nested type.

Instead, actually get the outer generic signature and filter requirements
that the outer signature satisfies.

Fixes <rdar://problem/53769896>, <https://bugs.swift.org/browse/SR-11221>.