SILGen: Add test to ensure we don't have a vtable entry for synthesized members of a final class
I believe this was fixed with https://github.com/apple/swift/pull/23932.

5.1 had an inconsistency here because the synthesized declaration was not
marked as final; we would sometimes check if a method itself was final,
other times we would check if both the method and the class were final.

Lucky it worked out so that the emitted vtable entry did not appear to
ever have been used.

Make sure this behavior doesn't change going forward with a test.