[ownership] When creating a new function declaration for deserialization, set the ownership eliminated flag.
I found this issue by inspection. I do not think it is actually possible to hit
this bug today since:

1. To hit the bug one would need to link in a definition of a function when we
do not have a pre-existing function. If we create the declaration first, then we
set the flag correctly when we deserialize the body of the function.

2. We today always set a callback on functions when we link in the body and
always strip ownership/set that flag. So we /should/ always set it correctly.

That being said, it is better to match what was actually serialized as close as
possible.