[ownership] Re-architect ownership pipeline so that we do not strip ossa from external functions.
Specifically, previously the plan was as follows:

```
-Onone:

* Run Diagnostic Passes.
* Serialize.
* Lower ownership.
* IRGen

-O:

* Run Diagnostic Passes
* Lower Ownership from all but transparent functions and register a
  deserialization notification to ensure any further non-transparent functions
  we linked in were stripped as well.
* Run Performance Passes Before Serialization.
* Serialize.
* Lower ownership from all functions and register a notification to ensure all
  further functions that are linked in are lowered on link.
* Run rest of performance passes.
* IRGen.
```

There is a problem with this setup though: what happens if we lower ownership
from a function, dce it, and then reload the function. Our declaration and new
definition will be inconsistent! After some thought, I realized that it really
did not make sense to just hide such inconsistency since the model in SIL is
that once ownership is removed, it can not be added back.

This commit works around this problem by creating 3 different styles of
ownership model eliminator, each varying what it will eliminate. Specifically:

1. Eliminate all + everything deserialized.
2. Eliminate non-external. Leave deserialized things alone.
3. Eliminate non-transparent, non-external only. Leave deserialized things alone.

We change the scheme such that we only lower external functions once we know we
will never DCE/relink any functions.

```
-Onone:

* Run Diagnostic Passes.
* Serialize.
* (1).
* IRGen

-O:

* Run Diagnostic Passes
* (2).
* Run Performance Passes Before Serialization.
* Serialize.
* (3).
* Run rest of performance passes.
* (1).
* IRGen.
```

When the option is disabled, we always lower ownership for local functions right
before mandatory inlining in the diagnostic passes, in the same place we used to
run an instance of (1). This is safe to do since if all functions before they
are serialized have ownership lowered, all serialized SIL should only have
lowered ownership implying we can never deserialize a function with ownership
not lowered, the previous state of the world. ☐.