Desugar protocol compositions in extended types.
Nested compositions are flattened before being added to the where clause.
Flattening is necessary as the first protocol (in canonical order) might be
somewhere deeply nested, and we need to take it out to save it as the extended
type.

Typealiases are resolved as one would expect.

Fixes SR-11227 and linked rdar://problem/53712389.