Sema: Refactor checkTypeModifyingDeclAttributes()
There were multiple calls to this function; all but one were not actually
needed.

Then for the last caller, instead of changing the VarDecl's interface type
to wrap it in a ReferenceStorageType, take the type as a parameter and
return a new type; the caller then sets the VarDecl's interface type
only once.