build-presets: Update/simplify incremental Xcode preset
This is our /only/ standard Xcode preset, and it's been disabled for a
long time. Since only a limited configuration is supported for
building with Xcode (pretty much just LLVM and Swift, with only a
macOS standard library), it doesn't make sense to base it on the
regular incremental preset. Pare it down to what's most important to
keep working in hopes of turning the bot back on.