[Constraint graph] Further subdivide "gathering kind" for gatherConstraints
Most clients of gatherConstraints() only care about the constraints
directly on the given type variable or other type variables equivalent
to it, while some also want to see constraints on fixed bindings as
well. Split "gathering kind" further so we do less work to retrieve
the constraints within an equivalence class for clients that only need
that information.