AST: Factor out common code into ModuleNameLookup::lookupInFileUnit()
Both UnqualifiedLookup and lookupVisibleDecls() had common logic to visit
the imports of a source file. Factor this out and simplify the interface
to lookupInModule().

Ideally we would get to a point where we only do unqualified lookups from
source files, so we can get rid of the other entry point altogether.
But for now there's no harm in keeping both.