[AutoDiff] Initial Forward Mode AD Support (#26057)
This PR extends on the previous work here on the `JVPEmitter` (https://github.com/apple/swift/pull/25954) to make it be able to generate SIL code for the `differential` function, as well as more correctly implementing the body of the JVP.'

generates a valid JVP and differential. I have yet to add support for control flow and generics. This PR also renames `PullbackInfo` to `LinearMap` info since it has a lot of tangent mapping helpers that `JVPEmitter` can use when creating the differential.

This is still early, so the JVP/Differential emission only runs with the flag `-Xllvm -run-jvp-generation` when running Swift. With it, you can differentiate basic Float functions with `+ - / *` with no `var`s