[semantic-arc-opts] Convert a piece of code to be an assert.
Ignoring bugs, any owned value that is used in a read only manner should be able
to be borrowed and used in such a place as well.

Originally this code (before refactoring) would assert, so we were checking that
this could not happen in our asserts builds. Due to refactoring the assert was
lost. This changes the code to have the assert again, this time explicitly
instead of inside the method.