[ModuleInterface] Don't print property wrapper storage vars
Previously, we would print the backing _val property for a given
property with a wrapper. This ended up printing incorrect output or
crashing when printing the initializer for these properties, because the
initializer consists of wholly-constructed calls to
MyWrapper.init(wrappedValue:). We don't need these storage vars, so just
omit them.

Fixes rdar://53555890 and rdar://53468607