[semantic-arc-opts] Add a worklist.
We first build up the worklist by applying our visitor to the CFG via a simple
instruction walk. During this walk, we know that we will only ever modify
instructions that are dominated by the visited instruction since we do not look
through basic block arguments.

Then once we have finished traversing the CFG, optimizing and gathering
opportunities, we drain the worklist. This is eventually where we should process
block arguments that we may be able to optimize.