Define SWIFT_REQUEST
Begin refactoring the request evaluator by swapping SWIFT_TYPEID for
SWIFT_REQUEST.  Introduce the Zone of the request as a formal parameter
to the macro, then re-expand the request macro to get the type info
back.

SWIFT_REQUEST will eventually grow to encompass more information about
requests as we seek to reduce the boilerplate involved in their
definitions.